Updated to v2.60.
