+++ FTIR_fittingtool_2_class.py
@@ -19,93 +19,10 @@ from sys import platform as _platform
 import ftir_sql_browser
 import guessnumbers
 import configparser
+from ColorTheme import color_theme
 
-__version__ = '2.59'
+__version__ = '2.60'
 __emailaddress__ = "pman3@uic.edu"
-
-
-class color_theme:
-    def __init__(self, theme):
-        self.theme = int(theme)
-        self.bg = ''
-        self.fg = ''
-        self.bg_toolbar = ''
-        self.bg_log = ''
-        self.facecolor = ''
-        self.warningcolor1 = ''
-        self.warningcolor2 = ''
-        self.warningcolor3 = ''
-
-        self.changetheme()
-
-    def changetheme(self):
-        if self.theme == 0:  # Dark: Default theme
-            self.bg = "#2b2b2b"
-            self.fg = "#a9b7c6"
-            self.bg_toolbar = '#262626'
-            self.bg_log = '#393c43'
-            self.facecolor = 'gainsboro'
-            self.warningcolor1 = 'red'
-            self.warningcolor2 = 'yellow'
-            self.warningcolor3 = 'royalblue'
-        elif self.theme == 1:  # Light
-            self.bg = "whitesmoke"
-            self.fg = "dimgrey"
-            self.bg_toolbar = 'gainsboro'
-            self.bg_log = 'darkgrey'
-            self.facecolor = 'white'
-            self.warningcolor1 = 'red'
-            self.warningcolor2 = 'yellow'
-            self.warningcolor3 = 'royalblue'
-        elif self.theme == 2:  # Royal
-            self.bg = "orangered"
-            self.fg = "white"
-            self.bg_toolbar = 'gold'
-            self.bg_log = 'gold'
-            self.facecolor = 'lemonchiffon'
-            self.warningcolor1 = 'red'
-            self.warningcolor2 = 'green'
-            self.warningcolor3 = 'royalblue'
-
-        elif self.theme == 3:  # Sky
-            self.bg = "lightskyblue"
-            self.fg = "white"
-            self.bg_toolbar = 'dodgerblue'
-            self.bg_log = 'cornflowerblue'
-            self.facecolor = 'lightcyan'
-            self.warningcolor1 = 'red'
-            self.warningcolor2 = 'yellow'
-            self.warningcolor3 = 'darkviolet'
-
-        elif self.theme == 4:  # Creamy
-            self.bg = "papayawhip"
-            self.fg = "gray"
-            self.bg_toolbar = 'orange'
-            self.bg_log = 'khaki'
-            self.facecolor = 'seashell'
-            self.warningcolor1 = 'red'
-            self.warningcolor2 = 'green'
-            self.warningcolor3 = 'royalblue'
-
-        elif self.theme == 5:  # Mystery
-            self.bg = "mediumpurple"
-            self.fg = "white"
-            self.bg_toolbar = 'darkviolet'
-            self.bg_log = 'mediumorchid'
-            self.facecolor = 'lavender'
-            self.warningcolor1 = 'lime'
-            self.warningcolor2 = 'orange'
-            self.warningcolor3 = 'deepskyblue'
-
-        elif self.theme == 6:  # Spring
-            self.bg = "greenyellow"
-            self.fg = "black"
-            self.bg_toolbar = 'limegreen'
-            self.bg_log = 'mediumseagreen'
-            self.facecolor = 'palegreen'
-            self.warningcolor1 = 'red'
-            self.warningcolor2 = 'orange'
-            self.warningcolor3 = 'royalblue'
 
 
 class FIT_FTIR:
@@ -1981,6 +1898,7 @@ class FTIR_fittingtool_GUI(Frame):
 
     def help(self):
         if self.needmorehelp == 0:
+            self.addlog('*' * 60)
             self.addlog("FTIR fitting tool, with customization of layer structures. v{}. ".format(__version__),
                         self.warningcolor3)
             self.addlog('Open a FTIR .csv file --> Customize your layer structure on the right. '
@@ -2067,6 +1985,10 @@ class FTIR_fittingtool_GUI(Frame):
                 helplines.insert(END, '\n   Ctrl+S: Show Transmissions using the input parameters.')
 
             helplines.insert(END, '\n\nUpdate Log:')
+            helplines.insert(END, '\nv. 2.60:')
+            helplines.insert(END, '\n   This update is focused on performance and stability improvements. ')
+            helplines.insert(END, '\n   A lot of features are packed in v2.5. v2.6 will be stable and with '
+                                  'less bugs. ')
             helplines.insert(END, '\nv. 2.59:')
             helplines.insert(END, '\n   Now one can calculate extinction coefficient k(k is not T-dependant.)')
             helplines.insert(END, '\n   Extinction coefficient k of ZnSe, ZnS, Ge and BaF2 '
@@ -2136,7 +2058,7 @@ class FTIR_fittingtool_GUI(Frame):
             helplines.insert(END, '\n   Added function to prevent two structures stack together. ')
             helplines.insert(END, '\n   Toolbar buttons are optimized to fit in more function buttons. ')
             helplines.insert(END, '\n   Now the newest added layer will shw on top instead of bottom.')
-            helplines.insert(END, '\nv2.40:')
+            helplines.insert(END, '\nv. 2.40:')
             helplines.insert(END, '\n   Added material data for ZnSe, BaF2, Ge and ZnS for FPI project.')
             helplines.insert(END, '\n   Added settings function.')
             helplines.insert(END, '\n   Added saveresult function.')
